Basic Main Road layout
Could probably stand to be longer or have more twists and turns in the story, but this'll do. No bezier smoothing yet, that'll come soon.

Todo: 1) the Train that will always perfectly hit just behind the player (unless they stop within view of the train tracks?) and 2) The end cinematics, then 3) audio