Suddenly, Sound Salad!
Added a bunch of sound and made my build for the Jam's end. Obviously there's a lot missing: the entities make no sound, nor does the train, nor do the screen effects. But that can all come with later versions, this is all I can do for my submission tonight.