Entity Watcher Panic
Only implemented for the Watchers (maybe I need a parent class for both entities to implement some of this stuff with) but it's spooky and maybe a little too loud but sounds freakin awesome

Also the most basic of Mix Bus setups came along with this. At some point I'll need to do a proper mix, but due to some audio programming tests I'm undertaking that may be a little while.