Keep driving after scene change trigger
Instead of coming to a stop (because that was audible), scene changes now change the car to Script control and just keep going forward. The transitions tend to be on straight roads, so this should be fine, but it _feels_ like it could be a problem if I had QA testers and such.