Placed Entities
The Main Road is now littered with enemies. Also added a line of dialogue explaining the train as you approach it, increased the player's max and starting health from 5 to 8, messed with Entity and Train Crossing Alarm, and slightly increased how much the player can see when going at full speed by moving the camera target a little further out.